update SDK version to 1.1.0 - add new domain apis - add new versions for commodity apis - fix schema for some objects

+++ src/api/types/email_validate_response_address_security.rs
@@ -3,7 +3,8 @@ pub use crate::prelude::*;
 #[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq)]
 pub struct EmailValidateResponseAddressSecurity {
     #[serde(default)]
-    pub threat_score: i64,
+    #[serde(with = "crate::core::number_serializers")]
+    pub threat_score: f64,
     #[serde(default)]
     pub is_tor: bool,
     #[serde(default)]
@@ -35,7 +36,7 @@ impl EmailValidateResponseAddressSecurity {
 #[derive(Clone, PartialEq, Default, Debug)]
 #[non_exhaustive]
 pub struct EmailValidateResponseAddressSecurityBuilder {
-    threat_score: Option<i64>,
+    threat_score: Option<f64>,
     is_tor: Option<bool>,
     is_proxy: Option<bool>,
     proxy_type: Option<String>,
@@ -49,7 +50,7 @@ pub struct EmailValidateResponseAddressSecurityBuilder {
 }
 
 impl EmailValidateResponseAddressSecurityBuilder {
-    pub fn threat_score(mut self, value: i64) -> Self {
+    pub fn threat_score(mut self, value: f64) -> Self {
         self.threat_score = Some(value);
         self
     }
